Refactor ingest_pipe module and adjust SRM service handler
