+++ infra/common/ingest_pipe/main.tf
@@ -40,35 +40,61 @@ resource "aws_iam_policy" "db_secret_access" {
 ################################################################################
 # Lambda for Sequence Run Manager event handling
 
-# Ingest Lambda Role
-resource "aws_iam_role" "ingest_lambda_role" {
-  name_prefix = "${var.service_id}_IngestLambdaRole"
-  path        = var.iam_path
+# Create the actual Lambda function using a 3rd party module
+module "ingest_lambda" {
+  source = "terraform-aws-modules/lambda/aws"
 
-  assume_role_policy = jsonencode({
-    Version = "2012-10-17"
-    Statement = [
-      {
-        Action = "sts:AssumeRole"
-        Effect = "Allow"
-        Principal = {
-          Service = "lambda.amazonaws.com"
-        }
-      }
-    ]
-  })
+  function_name = var.lambda_function_name
+  handler       = var.lambda_function_handler
+  runtime       = "python3.13"
+  publish       = true
+
+  source_path = var.lambda_source_paths
+
+  layers = var.lambda_layers
+
+  environment_variables = {
+    DB_SECRET_NAME = data.aws_secretsmanager_secret.db_secret.name
+  }
+
+  vpc_subnet_ids         = module.common.main_vpc_private_subnet_ids
+  vpc_security_group_ids = [module.common.orcahouse_db_sg_id[terraform.workspace]]
+  attach_network_policy  = true
+
 }
 
-# Attach the policy to the Lambda role
+# Ingest Lambda Role
+# resource "aws_iam_role" "ingest_lambda_role" {
+#   name_prefix = "${var.service_id}_IngestLambdaRole"
+#   path        = var.iam_path
+
+#   assume_role_policy = jsonencode({
+#     Version = "2012-10-17"
+#     Statement = [
+#       {
+#         Action = "sts:AssumeRole"
+#         Effect = "Allow"
+#         Principal = {
+#           Service = "lambda.amazonaws.com"
+#         }
+#       }
+#     ]
+#   })
+# }
+
+# Attach a policy to the Lambda role for Secrets Manager access
 resource "aws_iam_role_policy_attachment" "ingest_secrets_policy_attachment" {
-  role       = aws_iam_role.ingest_lambda_role.name
+  # role       = aws_iam_role.ingest_lambda_role.name
+  role       = module.ingest_lambda.lambda_role_name
   policy_arn = aws_iam_policy.db_secret_access.arn
 }
-# Attach VPC access policy
-resource "aws_iam_role_policy_attachment" "ingest_lambda_vpc_access" {
-  role       = aws_iam_role.ingest_lambda_role.name
-  policy_arn = module.common.lambda_vpc_access_policy_arn
-}
+# Attach VPC access policy 
+# TODO: handled by the lambda module
+# resource "aws_iam_role_policy_attachment" "ingest_lambda_vpc_access" {
+#   role       = aws_iam_role.ingest_lambda_role.name
+#   policy_arn = module.common.lambda_vpc_access_policy_arn
+# }
+
 # Attach VPC access policy restriction
 # https://docs.aws.amazon.com/lambda/latest/dg/configuration-vpc.html#configuration-vpc-best-practice
 resource "aws_iam_policy" "ingest_lambda_vpc_access_restriction" {
@@ -93,7 +119,8 @@ resource "aws_iam_policy" "ingest_lambda_vpc_access_restriction" {
         "Condition" : {
           "ArnEquals" : {
             "lambda:SourceFunctionArn" : [
-              aws_lambda_function.ingest_event_handler.arn
+              # aws_lambda_function.ingest_event_handler.arn
+              module.ingest_lambda.lambda_function_arn
             ]
           }
         }
@@ -102,49 +129,50 @@ resource "aws_iam_policy" "ingest_lambda_vpc_access_restriction" {
   })
 }
 resource "aws_iam_role_policy_attachment" "ingest_lambda_vpc_access_restriction" {
-  role       = aws_iam_role.ingest_lambda_role.name
+  # role       = aws_iam_role.ingest_lambda_role.name
+  role = module.ingest_lambda.lambda_role_name
   policy_arn = aws_iam_policy.ingest_lambda_vpc_access_restriction.arn
 }
 
 # Create package for Lambda function
-data "archive_file" "ingest_lambda_package" {
-  type        = "zip"
-  source_dir  = var.lambda_source_dir
-  output_path = var.lambda_artefact_out_path
+# data "archive_file" "ingest_lambda_package" {
+#   type        = "zip"
+#   source_dir  = var.lambda_source_dir
+#   output_path = var.lambda_artefact_out_path
 
-  excludes = [
-    "__pycache__",
-    "*.pyc",
-    "*.pyo",
-    "*.pyd"
-  ]
-}
+#   excludes = [
+#     "__pycache__",
+#     "*.pyc",
+#     "*.pyo",
+#     "*.pyd"
+#   ]
+# }
 
 # Ingest Lambda function resource
-resource "aws_lambda_function" "ingest_event_handler" {
-  filename      = data.archive_file.ingest_lambda_package.output_path
-  function_name = var.lambda_function_name
-  role          = aws_iam_role.ingest_lambda_role.arn
-  handler       = var.lambda_function_handler
-  runtime       = "python3.13"
-  timeout       = 30
-  memory_size   = 128
+# resource "aws_lambda_function" "ingest_event_handler" {
+#   filename      = data.archive_file.ingest_lambda_package.output_path
+#   function_name = var.lambda_function_name
+#   role          = aws_iam_role.ingest_lambda_role.arn
+#   handler       = var.lambda_function_handler
+#   runtime       = "python3.13"
+#   timeout       = 30
+#   memory_size   = 128
 
-  layers = var.lambda_layers
+#   layers = var.lambda_layers
 
-  source_code_hash = data.archive_file.ingest_lambda_package.output_base64sha256
+#   source_code_hash = data.archive_file.ingest_lambda_package.output_base64sha256
 
-  vpc_config {
-    subnet_ids         = module.common.main_vpc_private_subnet_ids
-    security_group_ids = [module.common.orcahouse_db_sg_id[terraform.workspace]]
-  }
+#   vpc_config {
+#     subnet_ids         = module.common.main_vpc_private_subnet_ids
+#     security_group_ids = [module.common.orcahouse_db_sg_id[terraform.workspace]]
+#   }
 
-  environment {
-    variables = {
-      DB_SECRET_NAME = data.aws_secretsmanager_secret.db_secret.name
-    }
-  }
-}
+#   environment {
+#     variables = {
+#       DB_SECRET_NAME = data.aws_secretsmanager_secret.db_secret.name
+#     }
+#   }
+# }
 
 
 ################################################################################
@@ -163,15 +191,18 @@ resource "aws_cloudwatch_event_target" "ingest_lambda" {
   target_id      = "SendToIngestLambda"
   event_bus_name = module.common.orcabus_bus_name
   rule           = aws_cloudwatch_event_rule.ingest_event_ingestion.name
-  arn            = aws_lambda_function.ingest_event_handler.arn
+  # arn            = aws_lambda_function.ingest_event_handler.arn
+  arn            = module.ingest_lambda.lambda_function_arn
 
-  depends_on = [aws_lambda_function.ingest_event_handler, aws_cloudwatch_event_rule.ingest_event_ingestion]
+  # depends_on = [aws_lambda_function.ingest_event_handler, aws_cloudwatch_event_rule.ingest_event_ingestion]
+  depends_on = [module.ingest_lambda, aws_cloudwatch_event_rule.ingest_event_ingestion]
 }
 
 resource "aws_lambda_permission" "ingest_event_allow_invoke" {
   statement_id  = "AllowExecutionFromEventBridgeRule"
   action        = "lambda:InvokeFunction"
-  function_name = aws_lambda_function.ingest_event_handler.function_name
+  # function_name = aws_lambda_function.ingest_event_handler.function_name
+  function_name = module.ingest_lambda.lambda_function_name
   principal     = "events.amazonaws.com"
   source_arn    = aws_cloudwatch_event_rule.ingest_event_ingestion.arn
 }
