+++ infra/service-event-ingestion/lambda/srm_sc_event_handler/srm_sc_event_handler.py
@@ -6,63 +6,26 @@ import psycopg2
 from psycopg2.extras import RealDictCursor
 from psycopg2.extensions import AsIs
 from os.path import join
+from utils import get_secret, get_db_connection, push_to_db
 
 
-TABLE_NAME = "psa.sequence_run_state_change_events"
-SRSC_DETAIL_TYPE = "SequenceRunStateChange"
-SRSC_EVENT_SOURCE = "orcabus.sequencerunmanager"
-RECORD_SOURCE = f"{SRSC_EVENT_SOURCE}:{SRSC_DETAIL_TYPE}"
 # Get the secret name from environment variables
 DB_SECRET_NAME = os.environ["DB_SECRET_NAME"]
 
-# SQL_INSERT = "INSERT INTO psa.fastq_list_row_change_events (%s) VALUES %s;"
+DB_SCHEMA = "psa"
+TABLE_NAME = "sequencerunmanager_sequencerunstatechange"
+TABLE = f"{DB_SCHEMA}.{TABLE_NAME}"
+DETAIL_TYPE = "SequenceRunStateChange"
+EVENT_SOURCE = "orcabus.sequencerunmanager"
+
 # Prevent inserts of the same event record multiple times
 # TODO: consider hashing the event values and only insert records that differ
 SQL_INSERT = f"INSERT INTO {TABLE_NAME} (%s) SELECT %s WHERE NOT EXISTS (SELECT 1 FROM {TABLE_NAME} WHERE event_id = %s);"
 
-
-# function to retrieve DB credetials
-def get_secret(secret_name):
-    """
-    Retrieve secret from AWS Secrets Manager
-    """
-    print("Retrieving DB credentials from Secrets Manager...")
-    session = boto3.session.Session()
-    client = boto3.client("secretsmanager")
-
-    try:
-        get_secret_value_response = client.get_secret_value(SecretId=secret_name)
-    except ClientError as e:
-        raise e
-    else:
-        if "SecretString" in get_secret_value_response:
-            secret = json.loads(get_secret_value_response["SecretString"])
-            return secret
-        else:
-            print("Secret not found!")
-
-
-# function to establish DB connection
-def get_db_connection(credentials):
-    print("Establishing connection to database...")
-    username = credentials.get("username")
-    password = credentials.get("password")
-    host = credentials.get("host")
-    port = credentials.get("port")
-    dbname = credentials.get("dbname")
-    print("Connecting to the database...")
-
-    conn = psycopg2.connect(
-        host=host, database=dbname, user=username, password=password, port=port
-    )
-    if conn:
-        print("Connection established!")
-    else:
-        print("Connection failed!")
-    return conn
-
-
-DB_CREDENTIALS = get_secret(DB_SECRET_NAME)
+# DB connection
+session = boto3.session.Session()
+secretsmanager_client = boto3.client("secretsmanager")
+DB_CREDENTIALS = get_secret(DB_SECRET_NAME, secretsmanager_client)
 DB_CONNECTION = get_db_connection(DB_CREDENTIALS)
 
 
@@ -71,7 +34,7 @@ def handler(event, context):
     print(f"Event: {event}")
     try:
         srsc_data = parse_event(event)
-        push_to_db(srsc_data)
+        push_to_db(DB_CONNECTION, SQL_INSERT, srsc_data)
 
         print("Returning results.")
         return {
@@ -134,7 +97,6 @@ def parse_event(event):
     end_time = detail.get("endTime")
     ss_name = detail.get("sampleSheetName")
 
-
     srsc_data = {
         "event_id": event_id,
         "event_time": event_time,
@@ -150,21 +112,6 @@ def parse_event(event):
     print(f"Extracted data: {srsc_data}")
 
     return srsc_data
-
-
-def push_to_db(data):
-    print("Pushing data to database...")
-    with DB_CONNECTION:
-        with DB_CONNECTION.cursor() as cur:
-            values = str(tuple(data.values()))[1:-1]  # strip off tuple brackets
-            sql = cur.mogrify(
-                SQL_INSERT,
-                (AsIs(",".join(data.keys())), AsIs(values), data["event_id"]),
-            )
-            print(f"SQL to execute: {sql}")
-            # cur.execute(sql)
-
-    print("Data pushed to database!")
 
 
 def test_case():
